refactor: 타입 오류 수정 및 상수 as const 선언
