+++ client/src/constants/CasperCustom/customStep.tsx
@@ -11,13 +11,13 @@ export const CUSTOM_STEP_OPTION = {
     FORM: "form",
     FINISHING: "finishing",
     FINISH: "finish",
-};
+} as const;
 export const CUSTOM_STEP_OPTION_ARRAY = [
     CUSTOM_STEP_OPTION.PROCESS,
     CUSTOM_STEP_OPTION.FORM,
     CUSTOM_STEP_OPTION.FINISHING,
     CUSTOM_STEP_OPTION.FINISH,
-];
+] as const;
 
 export const CUSTOM_STEP_HEADLINE = {
     [CUSTOM_STEP_OPTION.PROCESS]: {
@@ -34,6 +34,7 @@ export const CUSTOM_STEP_HEADLINE = {
             </>
         ),
     },
+    [CUSTOM_STEP_OPTION.FINISHING]: null,
     [CUSTOM_STEP_OPTION.FINISH]: {
         title: "친구에게 이벤트를 공유하고 당첨 확률을 높이세요!",
         description: "*당첨 결과는 N월 N일에 발표할게요",
